collision rewrite check desc
I initially had generic entity as an inherit.
Then I went to put it in the playerinfo class itself.

that was pretty dumb.
I took it out again. I also went to change the code a little so that now in CollisionResponse when you check you will only see a switch case and you only need to check the enum type of ThatEntity because the first entity is ALWAYS player. yay.